Refactors ResourcesNav to reduce rerenders.
refs #3045
